Production bug: Error when reading tracking.log input files from default interval.
We were using a datetime as the default interval end, so that the partition
could be made on the hour.  But the tracking.log processing requires dates,
not datetimes, for comparison with dates loaded from file names.

So, use the `date` parameter for the partition, and `interval' for the log
processing.